+++ src/main/scala/xiangshan/XSCore.scala
@@ -141,93 +141,12 @@ abstract class XSCoreBase(parentName:String = "Unknown")(implicit p: config.Para
   val frontend = LazyModule(new Frontend(parentName + "frontend_")(p))
   val csrOut = BundleBridgeSource(Some(() => new DistributedCSRIO()))
 
-  val wbArbiter = LazyModule(new WbArbiterWrapper(exuConfigs, NRIntWritePorts, NRFpWritePorts))
-  val intWbPorts = wbArbiter.intWbPorts
-  val fpWbPorts = wbArbiter.fpWbPorts
-
-  // TODO: better RS organization
-  // generate rs according to number of function units
-  require(exuParameters.JmpCnt == 1)
-  require(exuParameters.MduCnt <= exuParameters.AluCnt && exuParameters.MduCnt > 0)
-  require(exuParameters.FmiscCnt <= exuParameters.FmacCnt && exuParameters.FmiscCnt > 0)
-  require(exuParameters.LduCnt == 2 && exuParameters.StuCnt == 2)
-
-  // one RS every 2 MDUs
-  val schedulePorts = Seq(
-    // exuCfg, numDeq, intFastWakeupTarget, fpFastWakeupTarget
-    Seq(
-      (AluExeUnitCfg, exuParameters.AluCnt, Seq(AluExeUnitCfg, LdExeUnitCfg, StaExeUnitCfg), Seq()),
-      (MulDivExeUnitCfg, exuParameters.MduCnt, Seq(AluExeUnitCfg, MulDivExeUnitCfg), Seq()),
-      (JumpCSRExeUnitCfg, 1, Seq(), Seq()),
-      (LdExeUnitCfg, exuParameters.LduCnt, Seq(AluExeUnitCfg, LdExeUnitCfg), Seq()),
-      (StaExeUnitCfg, exuParameters.StuCnt, Seq(), Seq()),
-      (StdExeUnitCfg, exuParameters.StuCnt, Seq(), Seq())
-    ),
-    Seq(
-      (FmacExeUnitCfg, exuParameters.FmacCnt, Seq(), Seq(FmacExeUnitCfg, FmiscExeUnitCfg)),
-      (FmiscExeUnitCfg, exuParameters.FmiscCnt, Seq(), Seq())
-    )
-  )
-
-  // should do outer fast wakeup ports here
-  val otherFastPorts = schedulePorts.zipWithIndex.map { case (sche, i) =>
-    val otherCfg = schedulePorts.zipWithIndex.filter(_._2 != i).map(_._1).reduce(_ ++ _)
-    val outerPorts = sche.map(cfg => {
-      // exe units from this scheduler need fastUops from exeunits
-      val outerWakeupInSche = sche.filter(_._1.wakeupFromExu)
-      val intraIntScheOuter = outerWakeupInSche.filter(_._3.contains(cfg._1)).map(_._1)
-      val intraFpScheOuter = outerWakeupInSche.filter(_._4.contains(cfg._1)).map(_._1)
-      // exe units from other schedulers need fastUop from outside
-      val otherIntSource = otherCfg.filter(_._3.contains(cfg._1)).map(_._1)
-      val otherFpSource = otherCfg.filter(_._4.contains(cfg._1)).map(_._1)
-      val intSource = findInWbPorts(intWbPorts, intraIntScheOuter ++ otherIntSource)
-      val fpSource = findInWbPorts(fpWbPorts, intraFpScheOuter ++ otherFpSource)
-      getFastWakeupIndex(cfg._1, intSource, fpSource, intWbPorts.length).sorted
-    })
-    println(s"inter-scheduler wakeup sources for $i: $outerPorts")
-    outerPorts
-  }
-
-  // allow mdu and fmisc to have 2*numDeq enqueue ports
-  val intDpPorts = (0 until exuParameters.AluCnt).map(i => {
-    if (i < exuParameters.JmpCnt) Seq((0, i), (1, i), (2, i))
-    else if (i < 2 * exuParameters.MduCnt) Seq((0, i), (1, i))
-    else Seq((0, i))
-  })
-  val lsDpPorts = Seq(
-    Seq((3, 0)),
-    Seq((3, 1)),
-    Seq((4, 0)),
-    Seq((4, 1))
-  ) ++ (0 until exuParameters.StuCnt).map(i => Seq((5, i)))
-  val fpDpPorts = (0 until exuParameters.FmacCnt).map(i => {
-    if (i < 2 * exuParameters.FmiscCnt) Seq((0, i), (1, i))
-    else Seq((0, i))
-  })
-
-  val dispatchPorts = Seq(intDpPorts ++ lsDpPorts, fpDpPorts)
-
-  val outIntRfReadPorts = Seq(0, 0)
-  val outFpRfReadPorts = Seq(0, 2)
-  val hasIntRf = Seq(true, false)
-  val hasFpRf = Seq(false, true)
-  val exuBlocks = schedulePorts.zip(dispatchPorts).zip(otherFastPorts).zipWithIndex.map {
-    case (((sche, disp), other), i) =>
-      LazyModule(new ExuBlock(sche, disp, intWbPorts, fpWbPorts, other, outIntRfReadPorts(i), outFpRfReadPorts(i), hasIntRf(i), hasFpRf(i)))
-  }
-
+  val backendTop = LazyModule(new BackendTop)
   val memBlock = LazyModule(new MemBlock(parentName = parentName + "memBlock_")(p.alter((site, here, up) => {
     case XSCoreParamsKey => up(XSCoreParamsKey).copy(
-      IssQueSize = exuBlocks.head.scheduler.memRsEntries.max
+      IssQueSize = backendTop.exuBlocks.head.scheduler.memRsEntries.max
     )
   })))
-
-  val wb2Ctrl = LazyModule(new Wb2Ctrl(exuConfigs))
-  wb2Ctrl.addWritebackSink(exuBlocks :+ memBlock)
-  val dpExuConfigs = exuBlocks.flatMap(_.scheduler.dispatch2.map(_.configs))
-  val ctrlBlock = LazyModule(new CtrlBlock(dpExuConfigs))
-  val writebackSources = Seq(Seq(wb2Ctrl), Seq(wbArbiter))
-  writebackSources.foreach(s => ctrlBlock.addWritebackSink(s))
 }
 
 class XSCore(parentName:String = "Unknown")(implicit p: config.Parameters) extends XSCoreBase(parentName)
@@ -251,142 +170,64 @@ class XSCoreImp(parentName:String = "Unknown",outer: XSCoreBase) extends LazyMod
 
   println(s"FPGAPlatform:${env.FPGAPlatform} EnableDebug:${env.EnableDebug}")
 
-  val frontend = outer.frontend.module
-  val ctrlBlock = outer.ctrlBlock.module
-  val wb2Ctrl = outer.wb2Ctrl.module
-  val memBlock = outer.memBlock.module
-  val exuBlocks = outer.exuBlocks.map(_.module)
+  val backendTop = outer.backendTop.module
 
-  ctrlBlock.io.hartId := io.hartId
-  exuBlocks.foreach(_.io.hartId := io.hartId)
+  val frontend = outer.frontend.module
+  val memBlock = outer.memBlock.module
+
+  backendTop.io.hartId := io.hartId
   memBlock.io.hartId := io.hartId
-  outer.wbArbiter.module.io.hartId := io.hartId
   frontend.io.reset_vector := io.reset_vector
 
-  io.cpu_halt := ctrlBlock.io.cpu_halt
-
-  io.cpu_halt := ctrlBlock.io.cpu_halt
-
-  outer.wbArbiter.module.io.redirect <> ctrlBlock.io.redirect
-  val allWriteback = exuBlocks.flatMap(_.io.fuWriteback) ++ memBlock.io.writeback
-  require(exuConfigs.length == allWriteback.length, s"${exuConfigs.length} != ${allWriteback.length}")
-  outer.wbArbiter.module.io.in <> allWriteback
-  val rfWriteback = outer.wbArbiter.module.io.out
+  io.cpu_halt := backendTop.io.cpu_halt
 
   // memblock error exception writeback, 1 cycle after normal writeback
-  wb2Ctrl.io.s3_delayed_load_error <> memBlock.io.s3_delayed_load_error
-
-  wb2Ctrl.io.redirect <> ctrlBlock.io.redirect
-  outer.wb2Ctrl.generateWritebackIO()
+  backendTop.io.s3_delayed_load_error <> memBlock.io.s3_delayed_load_error
 
   io.beu_errors.icache <> frontend.io.error.toL1BusErrorUnitInfo()
   io.beu_errors.dcache <> memBlock.io.error.toL1BusErrorUnitInfo()
 
-  require(exuBlocks.count(_.fuConfigs.map(_._1).contains(JumpCSRExeUnitCfg)) == 1)
-  val csrFenceMod = exuBlocks.filter(_.fuConfigs.map(_._1).contains(JumpCSRExeUnitCfg)).head
-  val csrioIn = csrFenceMod.io.fuExtra.csrio.get
-  val fenceio = csrFenceMod.io.fuExtra.fenceio.get
+  val csrioIn = backendTop.io.csrio
+  val fenceio = backendTop.io.fenceio
 
-  frontend.io.backend <> ctrlBlock.io.frontend
+  frontend.io.backend <> backendTop.io.frontend
   frontend.io.sfence <> fenceio.sfence
   frontend.io.tlbCsr <> csrioIn.tlb
   frontend.io.csrCtrl <> csrioIn.customCtrl
   frontend.io.fencei := fenceio.fencei
 
-  ctrlBlock.io.csrCtrl <> csrioIn.customCtrl
-  val redirectBlocks = exuBlocks.reverse.filter(_.fuConfigs.map(_._1).map(_.hasRedirect).reduce(_ || _))
-  ctrlBlock.io.exuRedirect <> redirectBlocks.flatMap(_.io.fuExtra.exuRedirect)
-  ctrlBlock.io.stIn <> memBlock.io.stIn
-  ctrlBlock.io.memoryViolation <> memBlock.io.memoryViolation
-  exuBlocks.head.io.scheExtra.enqLsq.get <> memBlock.io.enqLsq
-  exuBlocks.foreach(b => {
-    b.io.scheExtra.lcommit := ctrlBlock.io.robio.lsq.lcommit
-    b.io.scheExtra.scommit := memBlock.io.sqDeq
-    b.io.scheExtra.lqCancelCnt := memBlock.io.lqCancelCnt
-    b.io.scheExtra.sqCancelCnt := memBlock.io.sqCancelCnt
-  })
-  val sourceModules = outer.writebackSources.map(_.map(_.module.asInstanceOf[HasWritebackSourceImp]))
-  outer.ctrlBlock.generateWritebackIO()
+  backendTop.io.stIn <> memBlock.io.stIn
+  backendTop.io.memoryViolation <> memBlock.io.memoryViolation
+  backendTop.io.enqLsq <> memBlock.io.enqLsq
+  backendTop.io.sqDeq := memBlock.io.sqDeq
+  backendTop.io.lqCancelCnt := memBlock.io.lqCancelCnt
+  backendTop.io.sqCancelCnt := memBlock.io.sqCancelCnt
 
-  val allFastUop = exuBlocks.flatMap(b => b.io.fastUopOut.dropRight(b.numOutFu)) ++ memBlock.io.otherFastWakeup
-  require(allFastUop.length == exuConfigs.length, s"${allFastUop.length} != ${exuConfigs.length}")
-  val intFastUop = allFastUop.zip(exuConfigs).filter(_._2.writeIntRf).map(_._1)
-  val fpFastUop = allFastUop.zip(exuConfigs).filter(_._2.writeFpRf).map(_._1)
-  val intFastUop1 = outer.wbArbiter.intConnections.map(c => intFastUop(c.head))
-  val fpFastUop1 = outer.wbArbiter.fpConnections.map(c => fpFastUop(c.head))
-  val allFastUop1 = intFastUop1 ++ fpFastUop1
+  backendTop.io.otherFastWakeup <> memBlock.io.otherFastWakeup
 
-  ctrlBlock.io.dispatch <> exuBlocks.flatMap(_.io.in)
-  ctrlBlock.io.rsReady := exuBlocks.flatMap(_.io.scheExtra.rsReady)
-  ctrlBlock.io.enqLsq <> memBlock.io.enqLsq
-  ctrlBlock.io.sqDeq := memBlock.io.sqDeq
-  ctrlBlock.io.lqCancelCnt := memBlock.io.lqCancelCnt
-  ctrlBlock.io.sqCancelCnt := memBlock.io.sqCancelCnt
-
-  exuBlocks(0).io.scheExtra.fpRfReadIn.get <> exuBlocks(1).io.scheExtra.fpRfReadOut.get
-  exuBlocks(0).io.scheExtra.fpStateReadIn.get <> exuBlocks(1).io.scheExtra.fpStateReadOut.get
-
-  for((c, e) <- ctrlBlock.io.ld_pc_read.zip(exuBlocks(0).io.issue.get)){
-    // read load pc at load s0
-    c.ptr := e.bits.uop.cf.ftqPtr
-    c.offset := e.bits.uop.cf.ftqOffset
-  }
   // return load pc at load s2
-  memBlock.io.loadPc <> VecInit(ctrlBlock.io.ld_pc_read.map(_.data))
-  memBlock.io.issue <> exuBlocks(0).io.issue.get
+  memBlock.io.loadPc <> VecInit(backendTop.io.ld_pc_read.map(_.data))
+  memBlock.io.issue <> backendTop.io.issue
+
   // By default, instructions do not have exceptions when they enter the function units.
   memBlock.io.issue.map(_.bits.uop.clearExceptions())
-  exuBlocks(0).io.scheExtra.loadFastMatch.get <> memBlock.io.loadFastMatch
-  exuBlocks(0).io.scheExtra.loadFastImm.get <> memBlock.io.loadFastImm
+  backendTop.io.loadFastMatch <> memBlock.io.loadFastMatch
+  backendTop.io.loadFastImm <> memBlock.io.loadFastImm
 
-  val stdIssue = exuBlocks(0).io.issue.get.takeRight(exuParameters.StuCnt)
-  exuBlocks.map(_.io).foreach { exu =>
-    exu.redirect <> ctrlBlock.io.redirect
-    exu.allocPregs <> ctrlBlock.io.allocPregs
-    exu.rfWriteback <> rfWriteback
-    exu.fastUopIn <> allFastUop1
-    exu.scheExtra.jumpPc <> ctrlBlock.io.jumpPc
-    exu.scheExtra.jalr_target <> ctrlBlock.io.jalr_target
-    exu.scheExtra.stIssuePtr <> memBlock.io.stIssuePtr
-    exu.scheExtra.debug_fp_rat <> ctrlBlock.io.debug_fp_rat
-    exu.scheExtra.debug_int_rat <> ctrlBlock.io.debug_int_rat
-    exu.scheExtra.memWaitUpdateReq.staIssue.zip(memBlock.io.stIn).foreach{case (sink, src) => {
-      sink.bits := src.bits
-      sink.valid := src.valid
-    }}
-    exu.scheExtra.memWaitUpdateReq.stdIssue.zip(stdIssue).foreach{case (sink, src) => {
-      sink.valid := src.valid
-      sink.bits := src.bits
-    }}
-  }
-  XSPerfHistogram("fastIn_count", PopCount(allFastUop1.map(_.valid)), true.B, 0, allFastUop1.length, 1)
-  XSPerfHistogram("wakeup_count", PopCount(rfWriteback.map(_.valid)), true.B, 0, rfWriteback.length, 1)
 
-  ctrlBlock.perfinfo.perfEventsEu0 := exuBlocks(0).getPerf.dropRight(outer.exuBlocks(0).scheduler.numRs)
-  ctrlBlock.perfinfo.perfEventsEu1 := exuBlocks(1).getPerf.dropRight(outer.exuBlocks(1).scheduler.numRs)
-  ctrlBlock.perfinfo.perfEventsRs  := outer.exuBlocks.flatMap(b => b.module.getPerf.takeRight(b.scheduler.numRs))
+  backendTop.io.stIssuePtr <> memBlock.io.stIssuePtr
+  backendTop.io.stIn := memBlock.io.stIn
+
 
   csrioIn.hartId <> io.hartId
   csrioIn.perf <> DontCare
-  csrioIn.perf.retiredInstr <> ctrlBlock.io.robio.toCSR.perfinfo.retiredInstr
-  csrioIn.perf.ctrlInfo <> ctrlBlock.io.perfInfo.ctrlInfo
   csrioIn.perf.memInfo <> memBlock.io.memInfo
   csrioIn.perf.frontendInfo <> frontend.io.frontendInfo
 
   csrioIn.perf.perfEventsFrontend <> frontend.getPerf
-  csrioIn.perf.perfEventsCtrl     <> ctrlBlock.getPerf
   csrioIn.perf.perfEventsLsu      <> memBlock.getPerf
   csrioIn.perf.perfEventsHc       <> io.perfEvents
 
-  csrioIn.fpu.fflags <> ctrlBlock.io.robio.toCSR.fflags
-  csrioIn.fpu.isIllegal := false.B
-  csrioIn.fpu.dirty_fs <> ctrlBlock.io.robio.toCSR.dirty_fs
-  csrioIn.fpu.frm <> exuBlocks(1).io.fuExtra.frm.get
-  csrioIn.exception <> ctrlBlock.io.robio.exception
-  csrioIn.isXRet <> ctrlBlock.io.robio.toCSR.isXRet
-  csrioIn.trapTarget <> ctrlBlock.io.robio.toCSR.trapTarget
-  csrioIn.interrupt <> ctrlBlock.io.robio.toCSR.intrBitSet
-  csrioIn.wfi_event <> ctrlBlock.io.robio.toCSR.wfiEvent
   csrioIn.memExceptionVAddr <> memBlock.io.lsqio.exceptionAddr.vaddr
 
   csrioIn.externalInterrupt.msip := outer.clint_int_sink.in.head._1(0)
@@ -403,13 +244,16 @@ class XSCoreImp(parentName:String = "Unknown",outer: XSCoreBase) extends LazyMod
   fenceio.sfence <> memBlock.io.sfence
   fenceio.sbuffer <> memBlock.io.fenceToSbuffer
 
-  memBlock.io.redirect <> ctrlBlock.io.redirect
-  memBlock.io.rsfeedback <> exuBlocks(0).io.scheExtra.feedback.get
+
+  memBlock.io.redirect <> backendTop.io.redirect
+  memBlock.io.rsfeedback <> backendTop.io.rsfeedback
   memBlock.io.csrCtrl <> csrioIn.customCtrl
   memBlock.io.tlbCsr <> csrioIn.tlb
-  memBlock.io.lsqio.rob <> ctrlBlock.io.robio.lsq
-  memBlock.io.lsqio.exceptionAddr.isStore := CommitType.lsInstIsStore(ctrlBlock.io.robio.exception.bits.uop.ctrl.commitType)
+
   memBlock.io.itlb <> frontend.io.ptw
+
+  memBlock.io.lsqio.rob <> backendTop.io.robio.lsq
+  memBlock.io.lsqio.exceptionAddr.isStore := CommitType.lsInstIsStore(backendTop.io.robio.exception.bits.uop.ctrl.commitType)
 
   // if l2 prefetcher use stream prefetch, it should be placed in XSCore
   io.l2_pf_enable := csrioIn.customCtrl.l2_pf_enable
@@ -425,12 +269,8 @@ class XSCoreImp(parentName:String = "Unknown",outer: XSCoreBase) extends LazyMod
     Seq(
       ModuleNode(memBlock),
       ResetGenNode(Seq(
-        ModuleNode(exuBlocks.head),
-        ResetGenNode(
-          exuBlocks.tail.map(m => ModuleNode(m)) :+ ModuleNode(outer.wbArbiter.module) :+ ModuleNode(wb2Ctrl)
-        ),
+        ModuleNode(backendTop),
         ResetGenNode(Seq(
-          ModuleNode(ctrlBlock),
           ResetGenNode(Seq(
             ModuleNode(frontend)
           ))
@@ -441,4 +281,7 @@ class XSCoreImp(parentName:String = "Unknown",outer: XSCoreBase) extends LazyMod
 
   ResetGen(resetTree, reset, !debugOpts.FPGAPlatform, Some(io.dfx_reset))
 
+  backendTop.io.dfx_reset := io.dfx_reset
+
+  backendTop.io.writeback <> memBlock.io.writeback
 }
